Pervadinta is GLang i MVPmm
Lyg tais veikia...
