+++ .idea/runConfigurations/Interactive.xml
@@ -1,7 +1,7 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Interactive" type="Application" factoryName="Application">
-    <option name="MAIN_CLASS_NAME" value="edu.ktu.glang.interpreter.GLangInterpreter" />
-    <module name="glang-interpreter" />
+    <option name="MAIN_CLASS_NAME" value="edu.ktu.MVPmm.interpreter.MVPmmInterpreter" />
+    <module name="MVPmm-interpreter" />
     <option name="PROGRAM_PARAMETERS" value="-i" />
     <extension name="coverage">
       <pattern>
