+++ src/main/java/edu/ktu/MVPmm/interpreter/MVPmmInterpreter.java
@@ -1,0 +1,139 @@
+package edu.ktu.MVPmm.interpreter;
+
+import edu.ktu.MVPmm.MVPmmLexer;
+import edu.ktu.MVPmm.MVPmmParser;
+import org.antlr.v4.runtime.CharStream;
+import org.antlr.v4.runtime.CharStreams;
+import org.antlr.v4.runtime.CommonTokenStream;
+import org.antlr.v4.runtime.misc.ParseCancellationException;
+import org.antlr.v4.runtime.tree.ParseTree;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.util.Objects;
+
+public class MVPmmInterpreter {
+    public static void main(String[] args) {
+        // Initialize variables to hold parsed arguments
+        String filename = null;
+        boolean isInteractiveMode = false;
+
+        // Loop through program arguments
+        for (int i = 0; i < args.length; i++) {
+            switch (args[i]) {
+                case "-f" -> {
+                    // If the -f flag is provided, check if there is a filename argument after it
+                    if (i + 1 < args.length) {
+                        // If there is a filename argument, store it and skip the next iteration of the loop
+                        filename = args[i + 1];
+                        i++;
+                    } else {
+                        // If there is no filename argument, print an error message and the help information and exit the program
+                        System.err.println("Error: Missing filename argument for -f flag.");
+                        printHelp();
+                        System.exit(1);
+                    }
+                }
+                case "-i" ->
+                    // If the -i flag is provided, enable interactive mode
+                        isInteractiveMode = true;
+                case "-h" -> {
+                    // If the -h flag is provided, print the help information and exit the program
+                    printHelp();
+                    System.exit(0);
+                }
+                default -> {
+                    // If an invalid argument is provided, print an error message and the help information and exit the program
+                    System.err.println("Error: Invalid argument: " + args[i]);
+                    printHelp();
+                    System.exit(1);
+                }
+            }
+        }
+
+        try {
+            if (isInteractiveMode) {
+                processInteractiveInput();
+            } else {
+                processFile(filename);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+            System.exit(1);
+        }
+    }
+
+    private static void printHelp() {
+        System.out.println("Usage: java ArgumentParser [-f filename] [-i] [-h]");
+        System.out.println("-f filename\tPass a file as an argument");
+        System.out.println("-i\t\tEnable interactive mode");
+        System.out.println("-h\t\tDisplay help information");
+    }
+
+    private static void processInteractiveInput() throws IOException {
+        //SymbolTable symbolTable = new SymbolTable();
+        MVPLangScope mvpLangScope = new MVPLangScope();
+        BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
+        String input = "";
+        while (true) {
+            System.out.print("> ");
+            String line = reader.readLine();
+            if (Objects.equals(line, "exit")) {
+                break;
+            }
+            input += line + "\n";
+            try {
+                String output = executeCode(mvpLangScope, CharStreams.fromString(input));
+                if (output != null) {
+                    input = "";
+                    if (!output.equals("")) {
+                        System.out.println(output);
+                    }
+                }
+            } catch (Exception e) {
+                e.printStackTrace();
+                System.out.println("<ERROR> " + e.getMessage());
+                input = "";
+            }
+        }
+    }
+
+    public static void processFile(String filename) {
+        //SymbolTable symbolTable = new SymbolTable();
+        MVPLangScope mvpLangScope = new MVPLangScope();
+        try {
+            String output = executeCode(mvpLangScope, CharStreams.fromFileName(filename));
+            System.out.println("<PROGRAM OUTPUT>");
+            System.out.println(output);
+        } catch (Exception e) {
+            e.printStackTrace();
+            System.out.println("<ERROR> " + e.getMessage());
+        }
+    }
+
+    public static String execute(String program) {
+        return executeCode(new MVPLangScope(), CharStreams.fromString(program));
+    }
+
+    private static String executeCode(MVPLangScope scope, CharStream input) {
+        MVPmmLexer lexer = new MVPmmLexer(input);
+        CommonTokenStream tokens = new CommonTokenStream(lexer);
+        MVPmmParser parser = new MVPmmParser(tokens);
+        parser.removeErrorListeners();
+        MVPmmErrorListener errorListener = new MVPmmErrorListener();
+        parser.addErrorListener(errorListener);
+
+        ParseTree tree = parser.program();
+
+        if (errorListener.isHasSyntaxError()) {
+            throw new ParseCancellationException(errorListener.getErrorMsg());
+        }
+        if (errorListener.isPartialTree()) {
+            return null;
+        }
+
+        InterpreterVisitor interpreter = new InterpreterVisitor(scope);
+        return (String) interpreter.visit(tree);
+    }
+}
