+++ src/main/java/edu/ktu/MVPmm/interpreter/exception/MVPReturnStatementException.java
@@ -1,0 +1,9 @@
+package edu.ktu.MVPmm.interpreter.exception;
+
+public class MVPReturnStatementException extends RuntimeException{
+    public Object object;
+    public MVPReturnStatementException(Object oobject){
+        super(String.format("Return was called."));
+        this.object = oobject;
+    }
+}
