+++ src/main/java/edu/ktu/MVPmm/interpreter/types/ArrayValue.java
@@ -1,0 +1,30 @@
+package edu.ktu.MVPmm.interpreter.types;
+
+public class ArrayValue {
+    private Object[] value;
+    private Type type;
+
+    public ArrayValue(Type type, Object[] value){
+        this.type = type;
+        this.value = value;
+    }
+
+    /*public ArrayValue(ArrayValue orgValue){
+        this.type = Type.createType(orgValue.getType().getBaseType());
+        this.value = orgValue.getValue();
+    }*/
+    public void setValue(Object[] newValue){
+        value = newValue;
+    }
+    public Type getType(){ return type;}
+    public Object[] getValue(){ return value;}
+
+    @Override
+    public String toString(){
+        return value.toString();
+    }
+
+    public String typeToString(){
+        return type.getBaseType().toString();
+    }
+}
