--- src/main/java/edu/ktu/glang/interpreter/Function.java
@@ -1,14 +1,0 @@
-package edu.ktu.glang.interpreter;
-
-import edu.ktu.glang.GLangParser;
-
-public class Function {
-    private final GLangParser.FunctionDeclarationContext context;
-    public Function(GLangParser.FunctionDeclarationContext ctx) {
-        this.context = ctx;
-    }
-
-    public GLangParser.FunctionDeclarationContext getContext(){
-        return context;
-    }
-}
