--- src/main/java/edu/ktu/glang/interpreter/SymbolTable.java
@@ -1,26 +1,0 @@
-package edu.ktu.glang.interpreter;
-
-import edu.ktu.glang.interpreter.types.Value;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class SymbolTable {
-    private final Map<String, Value> table;
-
-    public SymbolTable() {
-        table = new HashMap<>();
-    }
-
-    public void put(String name, Value value) {
-        table.put(name, value);
-    }
-
-    public Value get(String name) {
-        return table.get(name);
-    }
-
-    public boolean contains(String name) {
-        return table.containsKey(name);
-    }
-}
