--- src/main/java/edu/ktu/glang/interpreter/exception/MVPReturnStatementException.java
@@ -1,9 +1,0 @@
-package edu.ktu.glang.interpreter.exception;
-
-public class MVPReturnStatementException extends RuntimeException{
-    public Object object;
-    public MVPReturnStatementException(Object oobject){
-        super(String.format("Return was called."));
-        this.object = oobject;
-    }
-}
