--- src/main/java/edu/ktu/glang/interpreter/types/BooleanType.java
@@ -1,19 +1,0 @@
-package edu.ktu.glang.interpreter.types;
-
-public class BooleanType extends Type{
-
-    public BaseType baseType;
-    public BooleanType(){
-        baseType = BaseType.BOOLEAN;
-    }
-
-    @Override
-    public BaseType getBaseType(){
-        return baseType;
-    }
-
-    @Override
-    public boolean isNumericType(){
-        return false;
-    }
-}
