--- src/main/java/edu/ktu/glang/interpreter/types/StringType.java
@@ -1,19 +1,0 @@
-package edu.ktu.glang.interpreter.types;
-
-public class StringType extends Type {
-    public BaseType baseType;
-
-    public StringType(){
-        baseType = BaseType.STRING;
-    }
-
-    @Override
-    public BaseType getBaseType(){
-        return baseType;
-    }
-
-    @Override
-    public boolean isNumericType(){
-        return false;
-    }
-}
