+++ src/test/java/edu/ktu/MVPmm/interpreter/PrintTest.java
@@ -1,0 +1,23 @@
+package edu.ktu.MVPmm.interpreter;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class PrintTest {
+
+    @Test
+    void print_integer() {
+        String program = """
+                         print(5);                     
+                         """;
+
+        String expected = """
+                          5
+                          """;
+
+        String actual = MVPmmInterpreter.execute(program);
+
+        assertEquals(expected, actual);
+    }
+}
