+++ src/test/java/edu/ktu/MVPmm/interpreter/ScopeTest.java
@@ -1,0 +1,102 @@
+package edu.ktu.MVPmm.interpreter;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+public class ScopeTest {
+    @Test
+    void variable_in_scope(){
+        String program = """
+                         if(5 == 5){
+                         int a(5);
+                         print(a);
+                         }
+                         """;
+
+        String expected = """
+                          5
+                          """;
+
+        String actual = MVPmmInterpreter.execute(program);
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    void variable_out_of_scope(){
+        String program = """
+                         if(5 == 5){
+                         int a(5);
+                         }
+                         print(a);
+                         """;
+        assertThrows(RuntimeException.class,
+                () -> {
+                    MVPmmInterpreter.execute(program);
+                });
+    }
+
+    @Test
+    void variable_out_of_function_scope(){
+        String program = """
+                        int a(5);
+                        fun void testas <> {
+                        print(a);
+                        }
+                        testa<>;
+                         """;
+
+        assertThrows(RuntimeException.class,
+                () -> {
+                    MVPmmInterpreter.execute(program);
+                });
+    }
+
+    @Test
+    void variable_only_in_function_scope(){
+        String program = """
+                        fun void testas () {
+                        int a(5);
+                        }
+                        testas();
+                        print(a);
+                         """;
+
+        assertThrows(RuntimeException.class,
+                () -> {
+                    MVPmmInterpreter.execute(program);
+                });
+    }
+
+    @Test
+    void return_in_function(){
+        String program = """
+                        fun void testas <> {
+                        print(5);
+                        return;
+                        print(6);
+                        }
+                        testas<>
+                         """;
+
+        String expected = """
+                          5
+                          """;
+
+        String actual = MVPmmInterpreter.execute(program);
+        assertEquals(expected, actual);
+    }
+
+    @Test
+    void return_in_main_program(){
+        String program = """
+                        return 5;
+                         """;
+
+        assertThrows(RuntimeException.class,
+                () -> {
+                    MVPmmInterpreter.execute(program);
+                });
+    }
+}
