--- src/test/java/edu/ktu/glang/interpreter/VariablesDeclarationTest.java
@@ -1,68 +1,0 @@
-package edu.ktu.glang.interpreter;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-public class VariablesDeclarationTest {
-
-    @Test
-    void declare_and_print_variable() {
-        String program = """
-                         int a(5);            
-                         print(a);       
-                         """;
-
-        String expected = """
-                          5
-                          """;
-
-        String actual = GLangInterpreter.execute(program);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    void undeclared_variable_throws_exception() {
-        String program = """
-                         a(5);            
-                         print(a);       
-                         """;
-
-        assertThrows(RuntimeException.class,
-                () -> {
-                    GLangInterpreter.execute(program);
-                });
-    }
-
-    @Test
-    void assing_variable_and_print() {
-        String program = """
-                         int a(5);
-                         a(6);          
-                         print(a);       
-                         """;
-
-        String expected = """
-                          6
-                          """;
-
-        String actual = GLangInterpreter.execute(program);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    void assing_variable_with_type_mismatch() {
-        String program = """
-                         int a(5);
-                         a("abc");                  
-                         """;
-
-        assertThrows(ClassCastException.class,
-                () -> {
-                    GLangInterpreter.execute(program);
-                });
-    }
-}
